First commit. I've added plain magic bitboards (found in the tables.py file), fancy magic bitboards (found in the fancy_magics.py) and created a C# implementation for extra speed, found in the MagicBitboards.cs file.
I plan to include detailed commits into my work to revisit the history of this project, as well as include a GUI of some sorts with my program to allow me to interact with my own creation. I also plan on filtering out the Python files literring the workspace as they are only included for sake of translation to C# as Python is what I'm natively familiar with. It also works better for getting results from simple operations.