Made major changes to documentation and files. I added more to the documentation detailing how I achieved the various aspects of this program. The .md file on sliding and static pieces was finished, and I migrated the pawns section of the static pieces to a new file about pawns and why you shouldn't use a tablebase for them.
I also added major updates to the main Board.cs file to include making and unmaking moves involving the en-passant rule.

I am yet to include making and unmaking castling rights, and I face a weird issue where pawns are being deleted off the board for some strange reason. My two solutions to this are to spend an hour debugging my spaghetti code that made sense when I was sleep deprived, or an hour rewriting the code so that I have an easier time using it in the future.