Finally removed the .py files for magic move gen. I forgot to mention in the previous commit that I also refactored not just the MakeMove and UndoMove functions, but also how the constructor interprets FEN strings, opting to have an upside-down representation of the board as opposed to spamming ^56 everywhere to convert, what I call, a "bitboard square" to an "array index": index 8 on an array is index 48 on an array.
MakeMove and UndoMove both now work on a case-by-case interpretation of the type of move played, sectioning out the behaviour of the board depending on the move it interprets. This should hopefully make reading and sharing code much easier to be performed.

I'm quite proud of how far this project has come, and it's only been 2 weeks.

Who knows where we'll be in another 2 weeks?