Solved Perft(4) but having errors with Perft(4) I _finally_ managed to get a correct result for `Perft(4)` in my chess program, but running Perft(5) has led to me getting sent down a rabbit hole of problems.
Somehow, pieces are disappearing off the board and raising errors when trying to move ghost pieces that aren't there, yet when I play moves down to the point of the error, it works perfectly fine, proving there's an issue with my `MakeMove()` or `UndoMove()` code, but I can't tell which, if there's a problem there at all, not to mention as I write this, it's 4:40am in the morning and I have zero idea what is going on. Nothing makes sense anymore and I've expended all my energy on this project, neglecting some much needed sleep in the process.

Hopefully, after suggesting this to the Stockfish community, I can get some answers as to what my problem is.

Goodnight!