Updated pawn code using Disservin's Smallbrain I decided my pawn code needed some updating for two reasons: the first was that it was split into two functions and there was a third function with a Colour parameter that would run either of the two functions, depending on the enum in the parameter, and returned the result. This desperately needed a change, so with the influence of Smallbrain's code and martinn9078's help, I successfully got all the code working again.
I also learnt that you cannot have the ampersand operator and the true operator doing two different things, which seems like a bit of a weird feature but oh well.

The second reason was because my code didn't factor in en-passant and checkmasks. For example, in the FEN string "k7/8/8/1r2pP1K/8/8/8/8 w - e6 1 1", en-passant would be illegal, but my old code didn't even bother to factor that in. With this new code though, thanks again to Disservin for the Smallbrain sample, it correctly evaluates that en-passant would be illegal and therefore doesn't consider it when generating moves.

This chess program has taught me an insane amount about both chess itself and C#. I still am not even close to having an engine so I cannot wait to see where this project goes.