Implementing all the beta-testing changes back to main branch (#3)
## First commit to beta testing branch
I included this branch for snippets of code I want to publish for help with, but not affecting the content on the main branch.

##  Updated Bitboard.cs to use different PopLSB method
Instead of `bitboard ^= 1UL << LSB`, I used `bitboard &= bitboard - 1` to prevent the overhead of shifting a 1 to the LSB place.

##  Simple intermediary commit.
I'm currently trying to fix two prominent errors:
1. It will for some reason, try and move ghost pieces, then error because those pieces don't actually exit.
2. It will, during perft, delete the king bit off the board, which then raises errors when trying to generate king moves or castling moves, as the king isn't present for such a move.

This commit was made for meowmeree to explore deeper into my code.

Commits will continue onto the `beta-testing` branch until a major breakthrough is reached, in which it will be committed to the `main` branch, like all other major commits.

Apologies to my future self and to anyone curious enough to read these commit messages that I didn't keep the code more up-to-date. There were way too many errors that committing would update the repository to feature incorrect and error-prone code. Thank you for understanding.

## Finally got through all 6,500 test cases.
It took an incredibly long time, about 3 or so days of pure debugging, but I finally got through all the tests without encountering any errors.

The two errors were:

1. Castling: Something went incredibly wrong with castling rights where they were being incorrectly handled, causing rooks to be taken then magically spawn back in after making a castling move then unmaking that same castling move. The problem has now been solved thankfully, and I hope I never have to deal with this again. I also added a custom `Nibble` type that can be found in `Types/Nibble.cs`, which also, now, holds the `Bitboard` struct I use everywhere in my code.

I also plan on separating the `MoveGeneration.cs` into its appropriate sectioned files: one for magic bitboards, one for king and knight tables, etc.

2. Pinmasks and knights: The king would end up getting taken during Perft due to illegal knight moves that had to do with how I handled knights and pinmasks, which meant that a knight, being direcly pinned to the king, would be able to jump from the horizontal section of the pinmask to the vertical section of the pinmask, allowing the king to be taken. It turns out that, when pinned, knights absolutely cannot move at all, so to fix this error, I added a check to abort the function for generating knight moves early if the knight was found to be a member of the pinmask, and that solved the issue.

I'm currently working on filtering out the nearly 2,000 failed test cases and I am immensely proud of my program successfully completing 20,000 test cases (each test case has 6 depths and I did 4 of those, so ~6,500 x 4 = ~26,000). It's something I never thought was possible. From stressing over implementing magic bitboards to a program that can accurately traverse the game tree of Chess, this project has come so far. And I've only done move generation.

_Who knows where I'll be once I get an AI into this thing?_

## Updated pawn code using Disservin's Smallbrain
I decided my pawn code needed some updating for two reasons: the first was that it was split into two functions and there was a third function with a Colour parameter that would run either of the two functions, depending on the enum in the parameter, and returned the result. This desperately needed a change, so with the influence of Smallbrain's code and martinn9078's help, I successfully got all the code working again.

I also learnt that you cannot have the ampersand operator and the true operator doing two different things, which seems like a bit of a weird feature but oh well.

The second reason was because my code didn't factor in en-passant and checkmasks. For example, in the FEN string "k7/8/8/1r2pP1K/8/8/8/8 w - e6 1 1", en-passant would be illegal, but my old code didn't even bother to factor that in. With this new code though, thanks again to Disservin for the Smallbrain sample, it correctly evaluates that en-passant would be illegal and therefore doesn't consider it when generating moves.

This chess program has taught me an insane amount about both chess itself and C#. I still am not even close to having an engine so I cannot wait to see where this project goes.