Began solving issues with adding Chess960 support The first main issue arose when I attempted to blindly throw in Chess960 support, not realising everything else that I needed to do. I spent a few days searching for bugs in how castling was implemented, fixing the .RayBetween() function so that it used rook moves for two squares on the same line instead of bishop moves.
I added a new Castling960 move type, as well as an entirely new Board class called Board960 in Board960.cs that holds all the implementation for everything Chess960,  making the constructor for it use the same FEN string as classical chess, just with Chess960 castling rights.

There's still a few bugs that need to be solved. I'm testing the Kiwipete position on a reasonable depth, but currently getting issues with en-passant moves not being taken down correctly. This is a bit unexpected, because they work the same way as my Board.cs class does - the code was copied and pasted with just the Castling move type swapped with the Castling960 move type.

Either way, I've made this commit on a new branch, so I can continuously keep Chess960 support updated, without clogging the main branch with further commits and doing things more cleanly.