Passed Kiwipete test on depth 6 [Chess960] I finally corrected all the problems with my castling implementation to Chess960.
I remember there were two problems. One of them was issues with castling rights, so I decided to redesign castling rights, removing the need for the Nibble datatype, and created its own struct to be used in my code. This brought about a few problems, like castling rights not being disabled for the correct side, but those were easily traceable and fixable.

The second problem was pieces being deleted, which, again, was an issue with castling. Basically, in the Kiwipete position, after a series of moves where the knight took the rook on h8, castling rights were not correctly being disabled, meaning Black was still able to short castle, spawning the rook out of nowhere. Then, after undoing the short-castle move, the knight was replaced with a rook, then after undoing the knight move, it ended up with a Black rook being where the White knight was supposed to be.

After fixing these two issues, everything ran fine and it successfully passed the Kiwipete position on a depth of 6, concluding the implementation of Chess960 in my move generation.

This now concludes my journey of move generation. There is now literally nothing else I need to complete.

Besides the GUI. But we don't talk about that. That's not important.

I think.