Zobrist Support (#6)
#1 Began Zobrist Hashing and cleaning workspace
--------------------------------------------------------
Turns out I messed up a few things:

1. The 3-move repetition creating a draw is for POSITIONS and not MOVES. I was counting 3 repeated moves instead of 3 repeated positions, and to correct my mistake, I have to implement  Zobrist Hashing, which is what I've begun today in `Zobrist.cs`.

2. I never implemented a 50-move rule and never validated it, which is something I'm starting now. There will be a function returning a boolean if the current board state is violating the 50-move rule. It's also going to be changed to be implemented as a 100-move clock, because a "move" in this case consists of White moving a piece then Black making a piece - NOT only one side moving a piece.

3. There's no option for Zobrist hashing in the event of creating an engine where a transposition table would be a must. This has already been addressed above in number 1 - I have added `Zobrist.cs` with the help of Ciecke, creator of Stormphrax.

I also cleaned up the workspace, migrating the CastlingRights struct to the Chess.Castling namespace because it suits it more. There's also the Square and Squares struct where I used polymorphism to leverage implicit type conversions when handling squares, as enums do not natively provide this functionality.

#2 Zobrist hashing has been successfully implemented.
-------------------------------------------------------------
Turns out I was checking to see if there was four-fold repetition instead of the proper three-fold repetition with the line: `if (pastHashes[pastHash]++ == 3)`. The problem with this line is that

#3 Zobrist hashing, 50-move rule and the 3-fold rule
----------------------------------------------------------
A few problems occured with my previous code, so I fixed those bugs. They've already been addressed in the README and in the previous commit message (I think?), so I don't see a need to mention them a second (or third) time here.

Pull request #6 - the end of an era.