Fixed how castling squares are interpreted Turns out I missed out a letter on one line and that literally turned the entire project upside-down.
Thankfully, this should be the last of the last bugs in this thing, so now I don't have to worry about any of this.

If I ever have any future revisits to this project, I'll have probably added a way to input and output PGN data, since it's actually quite simple, surprisingly. All you need is to handle a few edge cases, run a test suite and carry on with your day. In fact, maybe I'll do it tomorrow when I have some free time. Either that or work on my engine.