Added PGN support and insufficient material. (#7)
Added PGN Support
-----------------------
It's been another month, and I've been tinkering with this project here and there. Safe to say, I think it's all done. It's all finally done. Maybe it needs an AI opponent that I'll package in with this thing, but for now, as far as I'm concerned, it's all done. There isn't really anything else move-generation-wise that needs to be implemented anymore. Everything's finished.

The PGN support was surprisingly really easy. I wrote the regex while in isolation at school, and copied it over to this project where I finished everything at home. Note that some PGNs might raise draw-related errors where generating legal moves in a state where the board is currently a draw WILL raise exceptions, but besides that, it correctly translated all the positions. I'll run another check to see if it translated the PGNs properly, or it's just faking. It seems to be identifying them correctly, and SAN to LAN conversion works properly, so as far as I'm concerned, nothing else needs to be changed.

I'll squash and merge this pull request, then finally say goodbye to this project. Nothing else (besides that damned documentation) needs to be completed anymore.

Thank you to everyone, once again. You don't understand how much finishing this project has had an impact on me. It's been amazing, making something I thought I never could. Struggling to understand magic bitboards, staring sideways at the screen when people threw around terms like "mailbox" and "zobrist hashes" - and to finally be able to understand it all: the feeling is magical.

So thank you, once more.